サービス層のロジックをユースケースへ移行
