+++ test/purchase_decision_service_test.dart
@@ -1,11 +1,11 @@
 import 'package:flutter_test/flutter_test.dart';
 import 'package:oouchi_stock/domain/entities/inventory.dart';
 import 'package:oouchi_stock/domain/entities/price_info.dart';
-import 'package:oouchi_stock/domain/services/purchase_decision_service.dart';
+import 'package:oouchi_stock/domain/usecases/purchase_decision.dart';
 import 'package:oouchi_stock/domain/repositories/buy_prediction_repository.dart';
 import 'package:oouchi_stock/domain/entities/buy_item.dart';
 import 'package:oouchi_stock/domain/usecases/add_prediction_item.dart';
-import 'package:oouchi_stock/domain/services/auto_prediction_list_service.dart';
+import 'package:oouchi_stock/domain/usecases/auto_add_prediction_item.dart';
 
 class _FakeRepo implements BuyPredictionRepository {
   final List<BuyItem> items = [];
@@ -26,8 +26,8 @@ class _FakeRepo implements BuyPredictionRepository {
 }
 
 void main() {
-  // 購入判定サービスのテスト
-  final service = PurchaseDecisionService(
+  // 購入判定ユースケースのテスト
+  final decision = PurchaseDecision(
     2,
     cautiousDays: 3,
     bestTimeDays: 3,
@@ -66,23 +66,23 @@ void main() {
       );
 
   test('在庫ゼロは緊急対応', () {
-    expect(service.decide(invWith(0), price()),
+    expect(decision(invWith(0), price()),
         PurchaseDecisionType.emergency);
   });
 
   test('残り日数少なら慎重対応', () {
-    expect(service.decide(invWith(1, consumption: 10), price()),
+    expect(decision(invWith(1, consumption: 10), price()),
         PurchaseDecisionType.cautious);
   });
 
   test('在庫十分でセール中はまとめ買いチャンス', () {
-    expect(service.decide(invWith(3), price(sale: 80, regular: 100)),
+    expect(decision(invWith(3), price(sale: 80, regular: 100)),
         PurchaseDecisionType.bulkOpportunity);
   });
 
   test('日数少かつ十分値引きなら最も買い時', () {
     expect(
-        service.decide(invWith(1, consumption: 10),
+        decision(invWith(1, consumption: 10),
             price(sale: 80, regular: 100)),
         PurchaseDecisionType.bestTime);
   });
@@ -90,17 +90,17 @@ void main() {
   // 予報リスト追加処理のテスト
   test('慎重判定で予報リストへ追加', () async {
     final repo = _FakeRepo();
-    final auto = AutoPredictionListService(
-        AddPredictionItem(repo), service);
-    await auto.process(invWith(1), price(sale: 150, regular: 100));
+    final auto = AutoAddPredictionItem(
+        AddPredictionItem(repo), decision);
+    await auto(invWith(1), price(sale: 150, regular: 100));
     expect(repo.items.length, 1);
   });
 
   test('緊急判定では予報リストへ追加しない', () async {
     final repo = _FakeRepo();
-    final auto = AutoPredictionListService(
-        AddPredictionItem(repo), service);
-    await auto.process(invWith(0), price());
+    final auto = AutoAddPredictionItem(
+        AddPredictionItem(repo), decision);
+    await auto(invWith(0), price());
     expect(repo.items.isEmpty, true);
   });
 }
